+++ docs/CHANGELOG.rst
@@ -39,6 +39,7 @@ Changed
   ``workflow-cutadapt-star-fc-quant-single``,
   ``workflow-cutadapt-star-fc-quant-wo-depletion-single``
 - Replace ANONYMOUS_USER_ID with ANONYMOUS_USER_NAME in settings.py
+- Add java memory settings in ``gatk-genomicsdb-import``
 
 Fixed
 -----